requirements.txt
alembic==1.11.1
altgraph==0.17.3
aniso8601==9.0.1
anyio==4.3.0
appscript==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asyncio==3.4.3
backoff==2.2.1
bandit==1.7.8
bcrypt==4.0.1
bleach==1.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
croniter==1.4.1
cryptography==38.0.1
dagit==1.3.13
dagster==1.3.13
dagster-cloud==1.3.13
dagster-cloud-cli==1.3.13
dagster-graphql==1.3.13
docstring-parser==0.15
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
Faker==18.4.0
fastapi==0.95.1
fsspec==2023.6.0
github3.py==4.0.1
gql==3.4.1
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.56.0
grpcio-health-checking==1.56.0
gTTS==2.3.2
h11==0.14.0
html5lib==0.9999999
httptools==0.5.0
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
JayDeBeApi==1.2.3
Jinja2==3.1.2
JPype1==1.4.0
kafka-python==2.0.2
lxml==4.9.1
macholib==1.16.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mock-open==1.4.0
msal==1.19.0
multidict==6.0.4
ntlm-auth==1.5.0
numpy==1.23.3
oauthlib==3.2.2
Office365-REST-Python-Client==2.3.14
openpyxl==3.0.10
packaging==23.1
pandas==1.4.4
passlib==1.7.4
pbr==6.0.0
pendulum==2.1.2
pex==2.1.137
pika==1.3.2
playsound==1.3.0
pluggy==1.2.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.2
psycopg2==2.9.9
psycopg2-binary==2.9.3
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.15
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==2.7.0
pymongo==3.12.3
pyodbc==4.0.34
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.4.2
rsa==4.9
schedule==1.1.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.11
starlette==0.26.1
starlette-graphene3==0.6.0
stevedore==5.2.0
superset-api-client==0.6.0
tabulate==0.9.0
tomli==2.0.1
toposort==1.10
tqdm==4.65.0
# Editable install with no version control (tutorial==0.0.0)
-e /Users/suraponkunsri/Documents/dagster/my-tutorial-project
typer==0.9.0
types-cryptography==3.3.23
typing_extensions==4.11.0
universal_pathlib==0.0.23
uritemplate==4.1.1
urljoin==1.0.0
urllib3==1.26.12
uvicorn==0.21.1
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
xlrd==2.0.1
xlwings==0.27.15
yarl==1.9.2
